reorganize singleton string type extensions
